prettifier/requirements.txt
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
certifi==2025.4.26
distro==1.9.0
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.5
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
